requirements.txt
fastapi==0.109.0
SQLAlchemy==2.0.25
pydantic==2.5.3
alembic==1.13.1
asyncpg==0.29.0
uvicorn==0.26.0
pydantic-settings==2.1.0
gunicorn
pytest==7.4.4
httpx==0.26.0
pytest-asyncio==0.23.3
pre-commit==3.6.0

redis==4.6.0
fastapi-cache2==0.2.1
aiocache==0.12.2
starlette==0.35.1
types-redis==4.6.0.20240106
celery==5.3.4
flower==2.0.0

gspread==6.0.1
